fix: Exclude data fixtures and factories from Symfony's DI
These files were trying to be included in the production environment which failed because the Composer packages weren't available (--no-dev). Not 100% sure what changed between Symfony 7.3 and 7.4 to cause this issue. If I ever require any controller arguments in the factories I will need to change this, maybe with

when@test:
    App\Factory\:
        resource: '../src/Factory/'

in the config/services.yaml file, or with manual configuration.